requirements.txt
annotated-types==0.7.0
anyio==4.12.1
browserdriver==0.0.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.13.2
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.3.5
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.7
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
openai==2.16.0
orjson==3.11.6
ormsgpack==1.12.2
packaging==25.0
pandas==3.0.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uuid-utils==0.14.0
wheel==0.46.3
xxhash==3.6.0
zstandard==0.25.0
